fix(sessions): commit archive state before worktree cleanup
